component-entrypoints
- `no-index-mark`: компонент перенесен в `no-index` (patch)
- `no-index`: проп closing заменен пропом type=open (major)
- `no-index`: добавлена точка входа `index.ts` (patch)
- `panel`: добавлена точка входа `index.ts` (patch)
- `plate`: добавлена точка входа `index.ts` (patch)
- `portal`: добавлена точка входа `index.ts` (patch)
- `price`: добавлена точка входа `index.ts` (patch)
