--- src/no-index-mark/__stories__/index.stories.tsx
@@ -1,27 +1,0 @@
-import { NoIndexMark } from '@sima-land/ui-nucleons/no-index-mark';
-
-export default {
-  title: 'service/NoIndexMark',
-  component: NoIndexMark,
-  parameters: {
-    layout: 'padded',
-  },
-};
-
-export function Primary() {
-  return (
-    <>
-      <NoIndexMark />
-      &quot;Кто б ни были входящие сюда,
-      <br />
-      Оставьте здесь надежду навсегда!&quot;
-      <br />
-      Написано над адскими вратами
-      <br />
-      Зловещими и черными чертами
-      <NoIndexMark closing />
-    </>
-  );
-}
-
-Primary.storyName = 'Простой пример';
