+++ src/no-index/__test__/index.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import { NoIndex } from '..';
+import { NoIndex, NoIndexMark } from '..';
 
 describe('NoIndex', () => {
   it('should renders correctly', () => {
@@ -27,5 +27,20 @@ describe('NoIndex', () => {
     expect(container.childNodes[2].childNodes).toHaveLength(1);
     expect(container.childNodes[2].childNodes[0] instanceof Comment).toBe(true);
     expect(container.childNodes[2].childNodes[0].textContent).toBe('/noindex');
+  });
+});
+
+describe('NoIndexMark', () => {
+  it('should renders correctly without prop', () => {
+    const { container, rerender } = render(<NoIndexMark />);
+    expect(container.innerHTML).toEqual('<span><!--noindex--></span>');
+
+    rerender(<NoIndexMark type='open' />);
+    expect(container.innerHTML).toEqual('<span><!--noindex--></span>');
+  });
+
+  it('should renders correctly with prop "closed"', () => {
+    const { container } = render(<NoIndexMark type='close' />);
+    expect(container.innerHTML).toEqual('<span><!--/noindex--></span>');
   });
 });
