+++ src/no-index/index.ts
@@ -1,0 +1,3 @@
+export type { NoIndexProps, NoIndexMarkProps } from './types';
+export { NoIndex } from './no-index';
+export { NoIndexMark } from './no-index-mark';
