--- src/no-index/index.tsx
@@ -1,22 +1,0 @@
-import { ReactNode } from 'react';
-import { NoIndexMark } from '../no-index-mark';
-
-export interface NoIndexProps {
-  /** Содержимое. */
-  children?: ReactNode;
-}
-
-/**
- * Неиндексируемый контент.
- * @param props Свойства.
- * @return Элемент.
- */
-export function NoIndex({ children }: NoIndexProps) {
-  return (
-    <>
-      <NoIndexMark />
-      {children}
-      <NoIndexMark closing />
-    </>
-  );
-}
