+++ src/no-index/types.ts
@@ -1,0 +1,10 @@
+import type { ReactNode } from 'react';
+
+export interface NoIndexMarkProps {
+  type?: 'open' | 'close';
+}
+
+export interface NoIndexProps {
+  /** Содержимое. */
+  children?: ReactNode;
+}
