+++ src/panel/index.ts
@@ -1,0 +1,2 @@
+export type { PanelType, PanelProps } from './types';
+export { Panel, PanelStyledMarkup } from './panel';
