+++ src/plate/index.ts
@@ -1,0 +1,2 @@
+export type { PlateProps } from './types';
+export { Plate } from './plate';
