+++ src/portal/index.ts
@@ -1,0 +1,2 @@
+export type { PortalProps } from './types';
+export { Portal } from './portal';
