+++ src/price/index.ts
@@ -1,0 +1,2 @@
+export type { PriceProps } from './types';
+export { Price } from './price';
